requirements.txt
numpy
pandas
requests
tqdm
geopandas
shapely
rasterio
pyproj
affine
matplotlib
python-dotenv